// src/scene/Enums/CubeEdge.ts
export enum CubeEdge {
  FrontTop = 'FRONT_TOP',
  FrontDown = 'FRONT_DOWN',
  FrontLeft = 'FRONT_LEFT',
  FrontRight = 'FRONT_RIGHT',
  TopLeft = 'TOP_LEFT',
  TopRight = 'TOP_RIGHT',
  DownLeft = 'DOWN_LEFT',
  DownRight = 'DOWN_RIGHT',
  BackTop = 'BACK_TOP',
  BackDown = 'BACK_DOWN',
  BackLeft = 'BACK_LEFT',
  BackRight = 'BACK_RIGHT',
}